.history/home/models_20231214122454.py
from django.db import models
from django.contrib.auth.models import User
from cloudinary.models import CloudinaryField

# Create your models here.

class Author*(models.Model):
    name = models.CharField(max)