.history/requirements_20231214141200.txt
PyJWT==2.8.0
asgiref==3.2.3
boto3==1.12.42
botocore==1.15.42
chardet==3.0.4
dj-database-url==0.5.0

django-allauth==0.41.0
django-countries==6.0
django-crispy-forms==1.8.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.5
oauthlib==3.1.0
Pillow==7.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sqlparse==0.3.0
stripe==2.42.0
urllib3==1.25.7